requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.4
pyyaml>=6.0.2

# Dev
python-multipart>=0.0.9
